+++ History.md
@@ -15,7 +15,7 @@
 
 * Allow `git+` URL schemes for npm dependencies #844
 
-* Expose options `disableOplog`, `pollingInterval`, and `pollingThrottle` to `Cursor.find` for tuning observe parameters on the server.
+* Expose options `disableOplog`, `pollingIntervalMs`, and `pollingThrottleMs` to `Cursor.find` for tuning observe parameters on the server.
 
 Patches contributed by GitHub users vereed, devgrok, ...